+++ lib/core/di/app_providers_module.dart
@@ -1,5 +1,8 @@
+import 'package:flutter_secure_storage/flutter_secure_storage.dart';
 import 'package:flutter_template/core/source/common/app_database.dart';
 import 'package:get_it/get_it.dart';
+import 'package:shared_preferences/shared_preferences.dart';
+import 'package:flutter_template/core/source/common/local_shared_preferences_storage.dart';
 
 class AppProvidersModule {
   AppProvidersModule._privateConstructor();
@@ -13,12 +16,29 @@ class AppProvidersModule {
     locator._setupModule();
     await locator.allReady();
   }
+
+  Future<void> setupProviders(GetIt locator) {
+    locator._setupProviders();
+    return locator.allReady();
+  }
 }
 
 extension _GetItDiModuleExtensions on GetIt {
   void _setupModule() {
     registerSingletonAsync(
       () => $FloorAppDatabase.databaseBuilder('app_database.db').build(),
+    );
+  }
+
+  void _setupProviders() {
+    registerLazySingleton(FlutterSecureStorage.new);
+    registerSingletonAsync(() => SharedPreferences.getInstance());
+
+    registerSingletonAsync(
+      () async => LocalSharedPreferencesStorage(
+        get(),
+        await getAsync<SharedPreferences>(),
+      ).init(),
     );
   }
 }
